+++ extensions/atlassian-data-center/src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as withQuery } from "./with-query";
+export { default as CacheActions } from "./cache-actions";
+export { default as SearchFilter } from "./search-filter";