+++ extensions/odesli/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Odesli Changelog
 
+## [Add history, favourites, and Windows support] - 2025-11-26
+
+- Added history and favourites functionality
+- Added Windows OS support
+
 ## [2023-11-16-1] - 2023-11-16
 
 ### Documentation